requirements.txt
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pypdf>=4.0.0
gradio>=4.0.0
